backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
asyncpg==0.29.0
databases[asyncpg]==0.9.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.0
numpy==1.26.4
scipy==1.13.0
shapely==2.0.4
geopandas==0.14.4
pykrige==1.7.1
websockets==12.0
python-dotenv==1.0.1
fastkml
lxml
openpyxl==3.1.5
